requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
black==23.11.0
build==0.10.0
CacheControl==0.13.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
crashtest==0.4.1
cryptography==41.0.2
decorator==5.1.1
distlib==0.3.7
Django==4.2.7
django-extensions==3.2.3
django-phonenumber-field==7.2.0
dulwich==0.21.5
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.12.2
flake8==6.1.0
idna==3.4
importlib-metadata==6.7.0
installer==0.7.0
ipython==8.18.1
isort==5.12.0
jaraco.classes==3.2.3
jedi==0.19.1
jeepney==0.8.0
jsonschema==4.17.3
keyring==24.2.0
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
phonenumbers==8.13.26
Pillow==10.1.0
pipenv==2023.8.26
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydot==1.4.2
pyflakes==3.1.0
Pygments==2.17.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.14.0
trove-classifiers==2023.5.24
typing_extensions==4.8.0
urllib3==1.26.16
virtualenv==20.24.3
wcwidth==0.2.12
zipp==3.15.0
